requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic>=2.7.4
python-dotenv==1.0.1
groq==1.2.0
langfuse==2.33.0
gunicorn
streamlit>=1.35.0
requests>=2.31.0
plotly>=5.20.0
pandas>=2.1.0
